requirements.txt
requests
PyYAML
Flask
